requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
transformers==4.38.2
python-dotenv==1.0.0 